+++ torchbearer/__init__.py
@@ -29,14 +29,11 @@ Utilities
 """
 
 from .version import __version__
-
-# Hack: Tensorboard and metrics need Y_PRED and Y_TRUE so define them here as a string
-Y_PRED = 'y_pred'
-Y_TRUE = 'y_true'
+from .bases import *
+from .state import *
 
 from . import metrics
 from . import callbacks
-from .state import *
 from .trial import *
 from .torchbearer import *
 from . import cv_utils