+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+### Changed
+- Y_PRED, Y_TRUE and X can now equivalently be accessed as PREDICTION, TARGET and INPUT respectively
+### Deprecated
+### Removed
+### Fixed
+- Fixed a bug where the LiveLossPlot callback would trigger an error if run and evaluate were called separately
+- Fixed a bug where state key errors would report to the wrong stack level
+- Fixed a bug where the user would wrongly get a state key error in some cases
+
 ## [0.2.5] - 2018-12-19
 ### Added
 - Added flag to replay to replay only a single batch per epoch